% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\name{Pokemon}
\alias{Pokemon}
\title{Pokemon}
\description{
Pokemon

Pokemon
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Pokemon$new()}}
\item \href{#method-print}{\code{Pokemon$print()}}
\item \href{#method-levelup}{\code{Pokemon$levelup()}}
\item \href{#method-damage}{\code{Pokemon$damage()}}
\item \href{#method-absorb}{\code{Pokemon$absorb()}}
\item \href{#method-validate_factor}{\code{Pokemon$validate_factor()}}
\item \href{#method-clone}{\code{Pokemon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$new(
  species = "Squirtle",
  type1 = "Water",
  type2 = NA,
  hp = 100,
  defense = 0,
  offense = 0,
  level = 1,
  turn = TRUE,
  happiness = 0,
  evo.stage = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evo.stage}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-levelup"></a>}}
\if{latex}{\out{\hypertarget{method-levelup}{}}}
\subsection{Method \code{levelup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$levelup(x = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-damage"></a>}}
\if{latex}{\out{\hypertarget{method-damage}{}}}
\subsection{Method \code{damage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$damage(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-absorb"></a>}}
\if{latex}{\out{\hypertarget{method-absorb}{}}}
\subsection{Method \code{absorb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$absorb(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_factor"></a>}}
\if{latex}{\out{\hypertarget{method-validate_factor}{}}}
\subsection{Method \code{validate_factor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$validate_factor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
